clean up old ethereum les wallet code
We couldn't build with Go 1.23 because a dependency was using
//go:linkname in a way that is now restricted.
